+++ mytown/Models/ProductAttributes.cs
@@ -10,7 +10,7 @@ namespace mytown.Models
         [Key]
         [Column("attribute_id")]
         [JsonPropertyName("attribute_id")]
-        public int AttributeId { get; set; }
+        public long AttributeId { get; set; }
 
         [Required]
         [StringLength(100)]
